moved stability functions to separate file
